Took into consideration if user spelt word wrong
Fixed pythonBackend so it searches for the closest word instead of asking first in the console, which caused an error. Also made the backend script return a tuple of the word and definition. Also got rid of sleep as I no longer needed it to test the thread. lbl_word now gets its text from the tuple.